More verbose documentation for code.
