+++ TechTest/AnyCompany.Service/.env.dev
@@ -1,2 +1,2 @@
 DATABASE_URL=Server=(localdb)\v11.0;Integrated Security=true;
-ROOT_QUERIES_DIR=Resources/Queries
+ROOT_QUERIES_DIR=./Resources/Queries
