+++ TechTest/AnyCompany.Service/Container/DefaultContainer.cs
@@ -31,7 +31,7 @@ namespace AnyCompany.Service.Container
             IOrdersFacade ordersFacade = null)
         {
             Config = config ?? new DefaultConfig(); 
-            Logger = new StdOutLogger(Console.OpenStandardOutput());
+            Logger = new StreamLogger(Console.OpenStandardOutput());
             Cache = cache ?? new SimpleInMemoryCache(); 
             
             DbWrapper = dbWrapper ?? new RelationalDatabaseWrapper(Config.DATABASE_URL, Cache, Config.ROOT_QUERIES_DIR, Logger);
