+++ TechTest/AnyCompany.Service/Facades/ITaxesFacade.cs
@@ -7,6 +7,12 @@ namespace AnyCompany.Service.Facades
     /// </summary>
     public interface ITaxesFacade 
     {
+        /// <summary>
+        /// Returns the tax by the specified type.
+        /// //TODO: Advanced filtering by more parameters, as country, state, etc.
+        /// </summary>
+        /// <param name="taxType">i.e 'VAT'</param>
+        /// <returns></returns>
         Task<decimal> GetTaxByType(string taxType);
     }
 }
